mysql: remove the connector libs from RPM
Due to licensing issues we currently do not want
to ship our version of Presto with mysql jar.
Removed the plugin installation from server rpm
and deleted the references for mysql connector
from our docs as well.

Raptor connector has been removed as well as it depends
on mysql jdbc connector.

Testing: manually verified that the jars are not present
in Presto installation.